+++ changelog.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 1.2
+
+* Added `unwrap_start_tag`, `unwrap_end_tag`, `unwrap_comment` to `XmlElement`.
+* Added `trim_text` function for use with `Iterator::map`.
+
 ## 1.1
 
 * Added `XmlElement::unwrap_text`